feat(wasm): wasm32-wasip1 运行时归档 —— #[path] 复用主运行时源码，一份源码两个目标
新 crate qi-runtime/wasm（独立 workspace，staticlib libqi_runtime_wasm.a）：字符串 / JSON /
列表 / 哈希表 / 字节切片 / 数学 / 时间 / 闭包 / 反射 / 正则 / 加密 / 压缩 / 向量 / 词法索引
全部用 #[path] 从 ../src/stdlib 原样编进来，模块树按主运行时同形摆放，源文件里的 crate::
引用不用改。只手写宿主强相关的几件：打印、分配、goroutine（单线程就地跑）、Future（同步）、
异常（wasi libc 没有 setjmp，只能 abort）。

为什么不给主 crate 加 feature：非可选依赖里 rusqlite（bundled sqlite）和 reqwest→rustls→
aws-lc-sys 两棵原生 C 库树的 build script 在 wasm32 上直接失败，跟 feature 开关无关。

主运行时顺带改三处：
- os_ffi：wasi 上 std::env::temp_dir / std::process::id 会 panic（panic=abort 直接带走
  整个程序），wasm 下给约定值 "/tmp" / 0。
- sync_ffi：qi_sync_mutex_lock / unlock 返回 i32 → i64，与注册表声明的 整数 对齐
  （x86_64 上 i32 返回高位是垃圾但没人读，wasm32 上签名不匹配直接链接期挂）。
- 其余七个文件是 cfg 门：网络 / 数据库 / 子进程 / GUI / MCP 在 wasm 目标下不编。

wasm最小/ 重命名为 wasm/。clippy 对 not_unsafe_ptr_arg_deref 与主运行时同一口径 crate 级放开
（几百个 extern "C" 函数是 #[path] 进来的，逐个标 unsafe 等于改主运行时公开签名）。

验证：qi/tests/wasm 差分回归 2/2（原生 vs wasmtime 逐字节一致）；示例 试炼.qi 在 wasmtime
与 node + wasi-shim（浏览器那条路）都跑通；原生侧 codegen 回归 30/30、qi-test 15/15。

+++ src/stdlib/compress_ffi.rs
@@ -12,7 +12,7 @@ use std::os::raw::c_char;
 
 /// 压缩文件（gzip）
 #[no_mangle]
-pub extern "C" fn qi_compress_gzip_file(source: *const c_char, dest: *const c_char) -> i32 {
+pub extern "C" fn qi_compress_gzip_file(source: *const c_char, dest: *const c_char) -> i64 {
     if source.is_null() || dest.is_null() {
         return -1;
     }
@@ -52,7 +52,7 @@ pub extern "C" fn qi_compress_gzip_file(source: *const c_char, dest: *const c_ch
 
 /// 解压文件（gzip）
 #[no_mangle]
-pub extern "C" fn qi_compress_gunzip_file(source: *const c_char, dest: *const c_char) -> i32 {
+pub extern "C" fn qi_compress_gunzip_file(source: *const c_char, dest: *const c_char) -> i64 {
     if source.is_null() || dest.is_null() {
         return -1;
     }
